rename state transition v3_11_2 -> v3_12_0
