+++ crates/contract/src/v3_12_0_state.rs
